requirements.txt
flask
matplotlib
numpy
scipy
pyserial